add some styles
