+++ package.json
@@ -30,6 +30,7 @@
   ],
   "dependencies": {
     "axios": "0.18.0",
+    "normalize.css": "8.0.1",
     "react": "16.8.3",
     "react-redux": "6.0.1",
     "react-scripts": "2.1.5",
