+++ src/index.js
@@ -3,4 +3,6 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { Root } from './Root';
 
+import './shared-styles/global-styles.css';
+
 ReactDOM.render(<Root />, document.getElementById('root'));
